Tabs into swap component
