+++ src/components/ConnectModal/connectModal.scss
@@ -1,0 +1,3 @@
+.modalConnect{
+
+}
