+++ src/views/Home/Home.vue
@@ -14,23 +14,13 @@
     <ConnectModal id="connectWalletModal"></ConnectModal>
 
     <div id="main-content">
-      <h1 id="text-title">
-        Swap your Synths
-      </h1>
-      <div class="buttons-mode">
-        <button v-on:click="showMint" v-bind:class="{activeButton: mintView}" >Mint</button>
-        <button v-on:click="showSwap" :class="{activeButton: swapView}" >Swap</button>
-        <button v-on:click="showBurn" :class="{activeButton: burnView}" >Burn</button>
-      </div>
+
 
       <!-- Wait until assets are loaded-->
+      <h1 id="text-title">Swap your Synths</h1>
       <b-container id="swap-window" v-if="initialized">
         <b-row>
-          <b-col>
-            <Mint v-if="mintView" />
-            <Swap v-if="swapView" />
-            <Burn v-if="burnView" />
-          </b-col>
+          <Swap/>
         </b-row>
       </b-container>
 
@@ -43,8 +33,6 @@
       <a href="https://github.com/block42-blockchain-company/thorchain-synthetics-swap"><BIconGithub/></a>
       <a href="https://twitter.com/block42_company"><BIconTwitter/></a>
     </div>
-
-
   </div>
 </template>
