+++ src/views/Home/home.ts
@@ -1,6 +1,4 @@
 
-import Mint from "@/components/Mint/Mint.vue"
-import Burn from "@/components/Burn/Burn.vue"
 import Swap from "@/components/Swap/Swap.vue"
 import {Component} from "vue-property-decorator";
 import Vue from 'vue'
@@ -10,8 +8,6 @@ import {BIconGithub, BIconTwitter} from "bootstrap-vue";
 
 @Component({
     components: {
-        Mint,
-        Burn,
         Swap,
         ConnectModal,
         BIconGithub,
@@ -25,16 +21,6 @@ import {BIconGithub, BIconTwitter} from "bootstrap-vue";
 })
 
 export default class Home extends Vue {
-    mintView: boolean = true;
-    swapView: boolean = false;
-    burnView: boolean = false;
-
-    reset(){
-        this.$store.commit("resetAssetInput")
-        this.mintView = false;
-        this.swapView = false;
-        this.burnView = false;
-    }
 
     beforeMount(){
         const apiUrl = 'https://testnet.midgard.thorchain.info/v2/pools';
@@ -44,20 +30,5 @@ export default class Home extends Vue {
                     return { "asset": asset.asset }
                 }));
             });
-    }
-
-    showMint(){
-        this.reset()
-        this.mintView = true;
-    }
-
-    showSwap(){
-        this.reset()
-        this.swapView = true;
-    }
-
-    showBurn(){
-        this.reset()
-        this.burnView = true;
     }
 }
